requirements.txt
Package                      Version
---------------------------- ---------
annotated-types              0.7.0
blinker                      1.9.0
cachetools                   5.5.2
certifi                      2025.6.15
charset-normalizer           3.4.2
click                        8.1.8
Flask                        3.1.1
flask-cors                   6.0.1
google-ai-generativelanguage 0.6.15
google-api-core              2.25.1
google-api-python-client     2.176.0
google-auth                  2.40.3
google-auth-httplib2         0.2.0
google-generativeai          0.8.5
googleapis-common-protos     1.70.0
grpcio                       1.73.1
grpcio-status                1.71.2
gTTS                         2.5.4
httplib2                     0.22.0
idna                         3.10
itsdangerous                 2.2.0
Jinja2                       3.1.6
MarkupSafe                   3.0.2
pip                          25.1.1
proto-plus                   1.26.1
protobuf                     5.29.5
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pydantic                     2.11.7
pydantic_core                2.33.2
pydub                        0.25.1
pyparsing                    3.2.3
python-dotenv                1.1.1
requests                     2.32.4
rsa                          4.9.1
SpeechRecognition            3.14.3
tqdm                         4.67.1
typing_extensions            4.14.1
typing-inspection            0.4.1
uritemplate                  4.2.0
urllib3                      2.5.0
Werkzeug                     3.1.3
